+++ tests/utils/tracking-code.value-object.util.spec.ts
@@ -1,33 +1,33 @@
-import { TrackingCodeValueObject } from '../../src/utils/tracking-code.value-object';
+import { TrackingCodeValueObject } from '../../lib/utils/tracking-code.value-object';
 
 describe('tracking.value-object', () => {
-	it('should be defined', ()=>{
+	it('should be defined', () => {
 		const track = TrackingCodeValueObject.create;
-		expect(track).toBeDefined()
-	})
+		expect(track).toBeDefined();
+	});
 
-	it('should generate a new track if not provide value', ()=>{
+	it('should generate a new track if not provide value', () => {
 		const track = TrackingCodeValueObject.create();
-		expect(track.isSuccess).toBeTruthy()
-	})
+		expect(track.isSuccess).toBeTruthy();
+	});
 
-	it('should fail if provide an invalid value', ()=>{
+	it('should fail if provide an invalid value', () => {
 		const track = TrackingCodeValueObject.create('XAB-AS909-055');
-		expect(track.isFailure).toBe(true)
-	})
+		expect(track.isFailure).toBe(true);
+	});
 
-	it('should create with provided value', ()=>{
+	it('should create with provided value', () => {
 		const track = TrackingCodeValueObject.create('B4V-X17423-3710');
-		expect(track.isSuccess).toBeTruthy()
-	})
+		expect(track.isSuccess).toBeTruthy();
+	});
 
-	it('should generate a random code', ()=>{
+	it('should generate a random code', () => {
 		const track = TrackingCodeValueObject.generate();
 		expect(track).toHaveLength(15);
-	})
+	});
 
-	it('should generate a random code', ()=>{
+	it('should generate a random code', () => {
 		const track = TrackingCodeValueObject.create().getResult();
 		expect(track.value).toHaveLength(15);
-	})
+	});
 });