+++ CHANGELOG.md
@@ -13,7 +13,7 @@ All notable changes to this project will be documented in this file.
 - ISBNCodeValueObject
 - UPCCodeValueObject
 
-### 2.5.0 - 2021-09-16
+### 2.5.0 - 2021-09-18
 
 ---
 ### Changed